backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
python-dotenv==1.0.1
aiohttp==3.11.12
pydantic==2.10.6
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.37
langchain-google-genai==2.0.9
langchain-chroma==0.2.2
langchain-huggingface==0.1.2
chromadb==0.6.3
python-multipart==0.0.20
aiofiles==24.1.0
sentence-transformers==3.4.1
pypdf==5.3.0
PyPDF2==3.0.1
python-magic==0.4.27
sqlalchemy==2.0.38
alembic==1.14.1
tqdm==4.67.1